[2022-11-30] view.html 안에 댓글UI 구현 기능은 미구현이며, blevel Long타입으로 변경

+++ src/main/java/com/library/dto/BoardRequestDto.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class BoardRequestDto {
 	private Long id;
-	private short blevel;
+	private Long blevel;			// Long 타입변경[2022-11-30]
 	private String title;
 	private String content;
 	private int readCnt;
